requirements.txt
asgiref==3.8.1
astroid==3.3.8
beautifulsoup4==4.13.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.7.0
diff-match-patch==20241021
dill==0.3.9
Django==5.1.5
django-bootstrap5==24.3
django-import-export==4.3.5
et_xmlfile==2.0.0
fonttools==4.56.0
isort==6.0.0
mccabe==0.7.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pycparser==2.22
pydyf==0.11.0
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.2
weasyprint==64.0
webencodings==0.5.1
zopfli==0.2.3.post1
